third_party/blink/web_tests/fast/forms/select/option-selecting-expected.txt

Test the behavior of selecting and deselecting options in <select> elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

ListBox with no size attribute specified.
PASS menuListBoxNoSize.selectedIndex is 1
PASS menuListBoxNoSize.options[1].selected is true
Setting menuListBoxNoSize.options[2].selected = true
Setting menuListBoxNoSize.options[2].selected = false
PASS menuListBoxNoSize.selectedIndex is 1
PASS menuListBoxNoSize.options[1].selected is true

ListBox with size=1 attribute.
PASS menuListBoxSize1.selectedIndex is 1
PASS menuListBoxSize1.options[1].selected is true
Setting menuListBoxSize1.options[2].selected = true
Setting menuListBoxSize1.options[2].selected = false
PASS menuListBoxSize1.selectedIndex is 1
PASS menuListBoxSize1.options[1].selected is true

ListBox with size=3 attribute.
PASS menuListBoxSize3.selectedIndex is -1
Setting menuListBoxSize3.options[1].selected = true
Setting menuListBoxSize3.options[1].selected = false
PASS menuListBoxSize3.selectedIndex is -1
PASS menuListBoxSize3.options[0].selected is false
PASS menuListBoxSize3.options[1].selected is false
Setting menuListBoxSize3.selectedIndex = 1
PASS menuListBoxSize3.selectedIndex is 1
PASS menuListBoxSize3.options[1].selected is true
Setting menuListBoxSize3.selectedIndex = -1
PASS menuListBoxSize3.selectedIndex is -1
PASS menuListBoxSize3.options[0].selected is false
PASS menuListBoxSize3.options[1].selected is false

ListBox with multiple attribute.
PASS multipleListBox.selectedIndex is -1
Setting multipleListBox.options[1].selected = true
Setting multipleListBox.options[1].selected = false
PASS multipleListBox.selectedIndex is -1
PASS multipleListBox.options[0].selected is false
PASS multipleListBox.options[1].selected is false
Setting multipleListBox.options[1].selected = true
Setting multipleListBox.options[2].selected = true
PASS multipleListBox.options[1].selected is true
PASS multipleListBox.options[2].selected is true
Unselecting selected first option:
PASS multipleListBox2.options[0].selected is true
PASS multipleListBox2.options[0].selected = false; multipleListBox2.options[0].selected is false
PASS successfullyParsed is true

TEST COMPLETE
